requirements.txt
streamlit
pandas
numpy
networkx
python-igraph
leidenalg
python-louvain
scikit-learn
matplotlib
